[MAMA] Add ability to suppress logging of payload load failure. (#306)
Some middlewares may attempt to load multiple payload bridges during
initialization, some of which may not always exist. This unfortunately
results in MAMA clients having to deal with spurious 'ERROR' logging and
may make them miss occasions when the logging actually indicates a
problem.

This changes adds support for configurable suppressing of this logging
using the following parameter:

    mama.payload.suppress_load_failure_logging

Note: this has only been enabled for the specific case of missing
libraries for the payload. Other logging relating to payload loading
failures remains 'ERROR' level, as they represent a real issue which
should not be ignored.

Tested using the results of the
MamaInternalTestC.loadPayloadBridgeInvalidName test, and a variety of
boolean values in the properties.

Signed-off By: Damian Maguire <dmaguire@velatt.com>