[MAMA] Add support for autoloading payload bridges from config. (#304)
Intended to resolve #145.

Adding support for automatic loading of payload bridges based on
properties defined in the mama.properties file. Payloads are loaded
during the call to mama_open, after the entitlements bridges and default
payload have been loaded, but before middleware bridges.

Payloads are loaded based on properties which match the following
pattern:

    mama.payload.autload.<name>=True|False

Where 'name' is the same value expected in a call to
mama_loadPayloadBridge. Payload loading can then easily be disabled by
changing the value of the property to a 'false' value.

Tested with a range of valid and invalid payload bridges and properties.

Signed-off by: Damian Maguire <damian.j.maguire@gmail.com>