Modified MAMA default log level to NORMAL (#329)
Has previously been WARN. It's unpleasant to have to set a log
level to WARN to print something by default. The onus should be
on the application to choose its default log level if NORMAL is
not efficient, but NORMAL should be the default. Otherwise
stats, versions etc have to declare themselves as warnings just
to get printed.

Signed-off-by: Frank Quinn <fquinn.ni@gmail.com>