Added initial docker files for performing builds
Going down the docker road allows us to provide packages outside of the
redhat mock ecosystem by leveraging FPM. This change involves:

1. A docker file to recreate each OS environment
2. A script for installing dependencies on a fresh system
3. A script for actually performing the package

Note that the docker build creates the build environment including
dependencies and builds the RPM. The docker run command simply packages
up the results and may be used in conjunction with shared volumes to
make the package accessible via the host. E.g.

    docker build -t omc7 -f release_scripts/Dockerfile . --build-arg IMAGE=centos:7
    docker run -e VERSION=6.2.3 -v release:/apps/release --rm omc7

Still a bit of work to do around "where to put things" but should mostly
be there.

Signed-off-by: Frank Quinn <fquinn@cascadium.io>