Started work on validating docker created packages
After creating packages using docker, it's possible to have issues in
actually installing them. This is the beginning of work to split docker
into a two part build to separate RPM generation and the final docker
image. It has the added benefit of greatly reducing the resulting docker
image size.

Signed-off-by: Frank Quinn <fquinn@cascadium.io>