Added new OpenMAMA Playback Bridge (ompb)
This bridge will help allow users to test their applications without having
to set up an out-of-process publisher to connect to. It can be built by adding
the following option to cmake:

    -DWITH_OMPB_BRIDGE=ON

And can then be used with `-m ompb` where the following mama.properties config
parameters are supported (given a transport name of 'pb'):

    mama.ompb.transport.pb.file = /path/to/playback.mplay
    mama.ompb.transport.pb.rate = 100000
    mama.ompb.transport.pb.on_completion = rewind

Note that you can also pass "raise" to the on_completion parameter to get the
bridge to raise a SIGINT on playback completion which can then be caught in
your application to clean up the test run.

Signed-off-by: Frank Quinn <fquinn@cascadium.io>