Added mama_getAllProperties and mama_getAvailableTransportNames
This is to make it easier to find properties matching patterns,
particularly in higher level languages which had no existing option
for doing this.

The mama_getAvailableTransportNames implementation is purely based
on the properties available in currently loaded configuration.

This change includes mama_getPropertiesAsString which will list
all properties in a single allocated string buffer for easy parsing
in higher level languages.

Signed-off-by: Frank Quinn <fquinn@cascadium.io>